@@ -28,10 +28,10 @@
 </template>
 
 <script>
-import Dropdown from "@/components/atoms/Dropdown";
-import ValidationMessage from "@/components/atoms/ValidationMessage";
-import Chip from "@/components/atoms/Chip";
-import Icon from "@/components/atoms/Icon";
+import Dropdown from "@/components/atoms/Dropdown.vue";
+import ValidationMessage from "@/components/atoms/ValidationMessage.vue";
+import Chip from "@/components/atoms/Chip.vue";
+import Icon from "@/components/atoms/Icon.vue";
 
 export default {
   name: "ChipBox",
